gcp-vqvae/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gcp-vqvae"
version = "0.1.0"
description = "GCP-VQVAE inference wrapper (encode/embed/decode)"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
  "torch",
  "numpy",
  "pyyaml",
  "python-box",
  "tqdm",
  "h5py",
  "transformers",
  "x-transformers",
  "vector-quantize-pytorch",
  "ndlinear",
  "graphein",
  "biopython",
  "torch-geometric",
  "torch-cluster",
  "torch-scatter",
  "torch-sparse",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["gcp_vqvae", "gcp_vqvae.*"]
